requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
discord==2.3.2
discord.py==2.5.2
distro==1.9.0
dotenv==0.9.9
elevenlabs==1.54.0
exceptiongroup==1.2.2
ffmpeg==1.4
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
multidict==6.1.0
openai==1.65.5
pip==25.0.1
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.8.2
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==15.0.1
wheel==0.45.1
yarl==1.18.3
deepgram-sdk==2.12.0
flask==3.1.0